+++ itachallenge-challenge/src/main/java/com/itachallenge/common/exception/ForbiddenException.java
@@ -1,0 +1,8 @@
+package com.itachallenge.common.exception;
+
+public class ForbiddenException extends RuntimeException {
+    public ForbiddenException(String message) {
+        super(message);
+    }
+}
+